requirements.txt
aiohttp==3.8.6
aiolimiter==1.1.0
aiosignal==1.3.1
annotated-types==0.6.0
async-timeout==4.0.3
asynch==0.2.2
attrs==23.1.0
beautifulsoup4==4.12.2
certifi==2023.7.22
charset-normalizer==3.2.0
ciso8601==2.3.0
clickhouse-cityhash==1.0.2.4
frozenlist==1.4.0
idna==3.4
leb128==1.0.5
lz4==4.3.2
multidict==6.0.4
pydantic==2.4.2
pydantic-settings==2.0.3
pydantic_core==2.10.1
python-dotenv==1.0.0
python-metar==1.4.0
pytz==2023.3.post1
requests==2.31.0
soupsieve==2.5
tqdm==4.66.1
typing_extensions==4.8.0
tzlocal==5.1
urllib3==2.0.4
yarl==1.9.2
zstd==1.5.5.1
